Fix to separate packages archives and deriveddata
