+++ sources/deriveddata/deriveddata.go
@@ -1,16 +1,14 @@
-package sweepcache
+package deriveddata
 
 import (
 	"os"
 	"os/exec"
 	"os/user"
 	"path/filepath"
-	"strconv"
 	"strings"
-	"time"
 )
 
-func cachedDerivedDataPaths() ([]string, error) {
+func derivedDataPaths() ([]string, error) {
 	xcodeBuildLocationStyle, err := exec.Command("defaults", "read", "com.apple.dt.Xcode", "IDEBuildLocationStyle").Output()
 	if err != nil {
 		xcodeBuildLocationStyle = []byte("Unique")
@@ -50,39 +48,4 @@ func cachedDerivedDataPaths() ([]string, error) {
 
 	return paths, nil
 
-}
-
-func cachedArchivesPath() string {
-
-	usr, _ := user.Current()
-	xcodePlistPath := strings.Replace("~/Library/Preferences/com.apple.dt.Xcode", "~", usr.HomeDir, 1)
-	archivesPath, err := exec.Command("defaults", "read", xcodePlistPath, "IDECustomDistributionArchivesLocation").Output()
-	if err != nil {
-		return strings.Replace("~/Library/Developer/Xcode/Archives", "~", usr.HomeDir, 1)
-	}
-	return string(archivesPath)
-
-}
-
-func checkExpired(dir string, expired time.Time) (bool, error) {
-
-	splited := strings.Split(dir, "-")
-	year, err := strconv.Atoi(splited[0])
-	if err != nil {
-		return false, err
-	}
-
-	month, err := strconv.Atoi(splited[1])
-	if err != nil {
-		return false, err
-	}
-
-	day, err := strconv.Atoi(splited[2])
-	if err != nil {
-		return false, err
-	}
-
-	date := time.Date(year, time.Month(month), day, 0, 0, 0, 0, time.UTC)
-
-	return date.Before(expired), nil
 }
